+++ Fuji/Source/Drivers/PC/MFSound_PC.cpp
@@ -361,7 +361,7 @@ MF_API void MFSound_SetMasterVolume(float volume)
 	pDSPrimaryBuffer->SetVolume(MFSoundInternal_GetDecibels(volume));
 }
 
-uint32 MFSound_GetPlayCursorInternal(MFVoice *pVoice, uint32 *pWriteCursor)
+size_t MFSound_GetPlayCursorInternal(MFVoice *pVoice, size_t *pWriteCursor)
 {
 	MFCALLSTACK;
